+++ pkg/connector/users.go
@@ -107,7 +107,7 @@ func (u *userBuilder) List(ctx context.Context, parentResourceID *v2.ResourceId,
 		rv = append(rv, ur)
 	}
 
-	token := prepareNextToken(page, uint(len(users)), total)
+	token := prepareNextToken(page, len(users), total)
 	nextPage, err := bag.NextToken(token)
 	if err != nil {
 		return nil, "", nil, fmt.Errorf("databricks-connector: failed to create next page token: %w", err)
